# ARROW-3321: [C++] Improve integer parsing performance