@kszucs 2018-09-24
That's an impressive speedup!